feat: finish form component
