+++ src/Search/Header/Form.js
@@ -5,33 +5,34 @@ import styled from "styled-components";
 
 import InputDepart from "../../Main/Header/Form/InputDepart";
 import InputComming from "../../Main/Header/Form/InputComming";
+import ButtonForward from "../../Main/Header/Form/ButtonForward";
+import ButtonReverse from "../../Main/Header/Form/ButtonReverse";
+import ButtonPassenger from "../../Main/Header/Form/ButtonPassenger";
+import ButtonSearch from "../../Main/Header/Form/ButtonSearch";
 
 const Form = styled.form`
-  background-color: rgba(214, 41, 194, 0.35);
+  padding: 2.5rem 0 2rem 0;
 `;
 
 const Places = styled.div`
-  background-color: rgba(214, 41, 194, 0.35);
   @media (min-width: 768px) {
     display: inline-block;
     width: 100%;
   }
   @media (min-width: 1200px) {
     display: inline-block;
-    width: 39%;
+    width: 40%;
   }
 `;
 
 const DateRange = styled.div`
-  background-color: rgba(214, 41, 194, 0.35);
   @media (min-width: 768px) {
     display: inline-block;
-    width: 49%;
+    width: 50%;
   }
 `;
 
 const Service = styled.div`
-  background-color: rgba(214, 41, 194, 0.35);
   @media (min-width: 768px) {
     display: inline-block;
     width: 100%;
@@ -43,43 +44,39 @@ const Service = styled.div`
 `;
 
 const Passenger = styled.div`
-  background-color: rgba(214, 41, 194, 0.35);
   @media (min-width: 768px) {
     display: inline-block;
-    width: 49%;
+    width: 50%;
   }
 `;
 
 //  удалить
 
 const Depart = styled.div`
-  background-color: rgba(214, 41, 194, 0.35);
+  position: relative;
   @media (min-width: 768px) {
     display: inline-block;
-    width: 49%;
+    width: 50%;
   }
 `;
 const Comming = styled.div`
-  background-color: rgba(214, 41, 194, 0.35);
   @media (min-width: 768px) {
     display: inline-block;
-    width: 49%;
+    width: 50%;
   }
 `;
 
 const Forward = styled.div`
-  background-color: rgba(214, 41, 194, 0.35);
   @media (min-width: 768px) {
     display: inline-block;
-    width: 49%;
+    width: 50%;
   }
 `;
 
 const Reverse = styled.div`
-  background-color: rgba(214, 41, 194, 0.35);
   @media (min-width: 768px) {
     display: inline-block;
-    width: 49%;
+    width: 50%;
   }
 `;
 
@@ -95,12 +92,20 @@ const Search = () => (
     </Places>
     <Service>
       <DateRange>
-        <Forward />
-        <Reverse>4</Reverse>
+        <Forward>
+          <ButtonForward />
+        </Forward>
+        <Reverse>
+          <ButtonReverse />
+        </Reverse>
       </DateRange>
       <Passenger>
-        <Forward>5</Forward>
-        <Reverse>6</Reverse>
+        <Forward>
+          <ButtonPassenger />
+        </Forward>
+        <Reverse>
+          <ButtonSearch />
+        </Reverse>
       </Passenger>
     </Service>
   </Form>
